Add Dark Mode, Scheduled Date and General Changes
- Dark Mode

Added a switch on the nav bar that allows you to switch to dark mode.
Changed every component to support the new dark mode.

- Scheduled Date

Extended the schema to support a due date on the TODO's and changed the frontend to support this new element.

- Other changed

Move the 'All Entries' view to be on a flex grid to improve legibility on narrow screens and to allow wide screens to fit TODO's properly.

Switch to Font Awesome icons for the buttons for a more consistent UI and to avoid compatability issues with emoji's.

Other UI changes such as adding labels to the dates.